🎨 :: Named select
